Data loading in old and new schema should be in same transaction. Currently there is a bug because these are in different transactions.
